BugFixes: OQL -> SQL uses wrong table for columns,           projection not supported in where/orderby,           OQL -> SQL does not support OR, () and static expr
in detail:

1)
i found this bug when using a oql like this:
   select t.FK1 from M2NTable t where t.FK2 = $1
(manually resolved M-N relation)
but it results in sql:
   select ID, FK1, FK2 from M2NTable, FK2Table where FK2Table.FK2=?
this is wrong because FK2 is in M2NTable.
Also the join isn't needed nor has it a corresponding where term ...

Solution: when building sql don't use FileDescriptor.clsDesc
(wich is the type of the column) but use the (new)
FileDescriptor.contClsDesc (containing class, parent).
Because i cannot detemine how to get the containing class i implement
this functionality: add contClsDesc to FieldDescriptor (+descendants) and
fill it when adding to ClassDescriptor.

Also fixed:
- respect fromClassAlias in checkField.
- add toString, totStringEx (recursive) to ParseTReeNode, Token for
  better debugging

3)
we need oql to support something like
  select t from jis.SomeTable t
    where t.SomeForeignKey.Name1 like $1
       or t.SomeForeignKey.Name2 like $2

which was wrong transcoded in several ways:
first there was added a join but when adding my where clause:
  ... where SomeTable.SomeForeignKey = Table2.ID
        and Table2.Name1 = ? or Table2.Name2 = ?
  which isn't right because it means
  ... where ( SomeTable.SomeForeignKey = Table2.ID
        and Table2.Name1 = ? ) or Table2.Name2 = ?
second the aproach to split the where expr at '?' and transcode the
terms with addCondition and concat it with "AND" wich results in
something like
  where SomeTable.SomeForeignKey = Table2.ID
        and Table2.Name1 = ? and or.Table2.Name2 = ?
the same problem occours when you use static expr like
  ... where SomeField = "abc"
in conjunction.

So i removed the addCondition from ParseTreeWalker and modified some
more methods to correctly build SQL expr from the tree and set it
direct with JDBCQueryExpression.setWhere()
Also make sure the where is enclosed in () when following a join
condition.

ToDo: implement the new method in HsqlQueryExpression
          encodeColumn( String tableName, String columnName )
      santiago@catnet.com.mx has offered to do this